product/requirements.txt
pymongo==3.9.0
botocore==1.15.31
shortuuid==1.0.1
Flask==1.0.2
pika==1.1.0
boto3==1.12.31
Flask_Cors==3.0.8
python-dotenv==0.12.0
redis==3.4.1
dnspython
